fix: keep interpreting after first return
Whenever a function with a return was called, the program stops
interpreting rest of the code. Fix it by taking the saved return value,
executing the inner block and get the new return value, then set the
interpreter's return value to the saved value. It's a bit confusing but
if it works, it works.